requirements.txt
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
deemix==3.6.6
deezer-py==1.3.7
Flask==2.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mutagen==1.45.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pycryptodomex==3.15.0
requests==2.28.1
tomli==2.0.1
typing-extensions==4.3.0
urllib3==1.26.9
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
